+++ Rebus/Config/RebusConfigurer.cs
@@ -340,7 +340,7 @@ namespace Rebus.Config
 
             if (!_injectionist.Has<ITransport>())
             {
-                throw new Rebus.Exceptions.ConfigurationErrorsException(
+                throw new Rebus.Exceptions.RebusConfigurationException(
                     "No transport has been configured! You need to call .Transport(t => t.Use***) in order" +
                     " to select which kind of queueing system you want to use to transport messages. If" +
                     " you want something lightweight (possibly for testing?) you can use .Transport(t => t.UseInMemoryTransport(...))");