+++ Rebus/Pipeline/IncomingStepContext.cs
@@ -14,6 +14,8 @@ namespace Rebus.Pipeline
         public IncomingStepContext(TransportMessage message, ITransactionContext transactionContext)
         {
             Save(message);
+            Save(new OriginalTransportMessage(message));
+
             Save(transactionContext);
             
             transactionContext.Items[StepContextKey] = this;