+++ Rebus/Sagas/ICorrelationConfig.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq.Expressions;
+using Rebus.Pipeline;
 
 namespace Rebus.Sagas
 {
@@ -22,5 +23,27 @@ namespace Rebus.Sagas
         /// <param name="messageValueExtractorFunction">Configures a function to extract a value from the message. Since this is just a function, it may contain logic that e.g. concatenates fields, calls other functions, etc.</param>
         /// <param name="sagaDataValueExpression">Configures an expression, which will be used when querying the chosen <see cref="ISagaStorage"/> - since this is an expression, it must point to a simple property of the relevant <typeparamref name="TSagaData"/>.</param>
         void Correlate<TMessage>(Func<TMessage, object> messageValueExtractorFunction, Expression<Func<TSagaData, object>> sagaDataValueExpression);
+
+        /// <summary>
+        /// Correlates an incoming message of type <typeparamref name="TMessage"/> using the header with the given <paramref name="headerKey"/>. The value will be used when looking up a saga data instance using the specified <paramref name="sagaDataValueExpression"/>.
+        /// You could for example do something like this:
+        /// <code>
+        /// config.CorrelateHeader&lt;TradeApproved&gt;("trade-corr-id", d => d.TradeId);
+        /// </code>
+        /// to look up a saga instance by the "TradeId" field, querying by the value of the "trade-corr-id" header of the incoming "TradeApproved" message.
+        /// </summary>
+        /// <typeparam name="TMessage">Specifies the message type to configure a correlation for</typeparam>
+        /// <param name="headerKey">Configures a header key which will be extracted from the incoming message</param>
+        /// <param name="sagaDataValueExpression">Configures an expression, which will be used when querying the chosen <see cref="ISagaStorage"/> - since this is an expression, it must point to a simple property of the relevant <typeparamref name="TSagaData"/>.</param>
+        void CorrelateHeader<TMessage>(string headerKey, Expression<Func<TSagaData, object>> sagaDataValueExpression);
+
+        /// <summary>
+        /// Correlates an incoming message of type <typeparamref name="TMessage"/> using the message context to get a value (e.g. by selecting certain headers, combining them, etc)
+        ///  The value will be used when looking up a saga data instance using the specified <paramref name="sagaDataValueExpression"/>.
+        /// <typeparam name="TMessage">Specifies the message type to configure a correlation for</typeparam>
+        /// <param name="contextValueExtractorFunction">Configures a function that can extract a value from the current <see cref="IMessageContext"/></param>
+        /// <param name="sagaDataValueExpression">Configures an expression, which will be used when querying the chosen <see cref="ISagaStorage"/> - since this is an expression, it must point to a simple property of the relevant <typeparamref name="TSagaData"/>.</param>
+        /// </summary>
+        void CorrelateContext<TMessage>(Func<IMessageContext, object> contextValueExtractorFunction, Expression<Func<TSagaData, object>> sagaDataValueExpression);
     }
 }