+++ Rebus/Sagas/LoadSagaDataStep.cs
@@ -9,6 +9,7 @@ using Rebus.Logging;
 using Rebus.Messages;
 using Rebus.Pipeline;
 using Rebus.Pipeline.Receive;
+using Rebus.Transport;
 
 namespace Rebus.Sagas
 {
@@ -72,6 +73,7 @@ Afterwards, all the created/loaded saga data is updated appropriately.")]
 
             var message = context.Load<Message>();
             var label = message.GetMessageLabel();
+            var transactionContext = context.Load<ITransactionContext>();
 
             var body = message.Body;
 
@@ -82,7 +84,7 @@ Afterwards, all the created/loaded saga data is updated appropriately.")]
             // and then we process them
             foreach (var sagaInvoker in handlerInvokersForSagas)
             {
-                await TryMountSagaDataOnInvoker(sagaInvoker, body, label, loadedSagaData, newlyCreatedSagaData);
+                await TryMountSagaDataOnInvoker(sagaInvoker, body, label, loadedSagaData, newlyCreatedSagaData, transactionContext);
             }
 
             // invoke the rest of the pipeline (most likely also dispatching the incoming message to the now-ready saga handlers)
@@ -109,7 +111,7 @@ Afterwards, all the created/loaded saga data is updated appropriately.")]
             }
         }
 
-        async Task TryMountSagaDataOnInvoker(HandlerInvoker sagaInvoker, object body, string label, List<RelevantSagaInfo> loadedSagaData, List<RelevantSagaInfo> newlyCreatedSagaData)
+        async Task TryMountSagaDataOnInvoker(HandlerInvoker sagaInvoker, object body, string label, List<RelevantSagaInfo> loadedSagaData, List<RelevantSagaInfo> newlyCreatedSagaData, ITransactionContext transactionContext)
         {
             var foundExistingSagaData = false;
 
@@ -118,7 +120,7 @@ Afterwards, all the created/loaded saga data is updated appropriately.")]
 
             foreach (var correlationProperty in correlationPropertiesRelevantForMessage)
             {
-                var valueFromMessage = correlationProperty.ValueFromMessage(body);
+                var valueFromMessage = correlationProperty.ValueFromMessage(new MessageContext(transactionContext), body);
                 var sagaData = await _sagaStorage.Find(sagaInvoker.Saga.GetSagaDataType(), correlationProperty.PropertyName, valueFromMessage);
 
                 if (sagaData == null) continue;
@@ -143,7 +145,7 @@ Afterwards, all the created/loaded saga data is updated appropriately.")]
                     // if there's exacly one correlation property that points to a property on the saga data, we can set it
                     if (correlationPropertiesRelevantForMessage.Length == 1)
                     {
-                        TrySetCorrelationPropertyValue(newSagaData, correlationPropertiesRelevantForMessage[0], body);
+                        TrySetCorrelationPropertyValue(newSagaData, correlationPropertiesRelevantForMessage[0], body, transactionContext);
                     }
 
                     sagaInvoker.SetSagaData(newSagaData);
@@ -160,7 +162,7 @@ Afterwards, all the created/loaded saga data is updated appropriately.")]
             }
         }
 
-        static void TrySetCorrelationPropertyValue(ISagaData newSagaData, CorrelationProperty correlationProperty, object body)
+        static void TrySetCorrelationPropertyValue(ISagaData newSagaData, CorrelationProperty correlationProperty, object body, ITransactionContext transactionContext)
         {
             try
             {
@@ -170,7 +172,7 @@ Afterwards, all the created/loaded saga data is updated appropriately.")]
 
                 if (correlationPropertyInfo == null) return;
 
-                var valueFromMessage = correlationProperty.ValueFromMessage(body);
+                var valueFromMessage = correlationProperty.ValueFromMessage(new MessageContext(transactionContext), body);
 
                 correlationPropertyInfo.SetValue(newSagaData, valueFromMessage);
             }